Rename componentsin StreamField using CamelCase convention
